+++ crates/tauri/src/manager/mod.rs
@@ -31,6 +31,9 @@ use crate::{
   Assets, Context, DebugAppIcon, EventName, Pattern, Runtime, StateManager, Webview, Window,
 };
 
+#[cfg(any(target_os = "macos", target_os = "ios"))]
+use crate::app::OnWebContentProcessTerminate;
+
 #[cfg(desktop)]
 mod menu;
 #[cfg(all(desktop, feature = "tray-icon"))]
@@ -251,6 +254,9 @@ impl<R: Runtime> AppManager<R> {
     plugins: PluginStore<R>,
     invoke_handler: Box<InvokeHandler<R>>,
     on_page_load: Option<Arc<OnPageLoad<R>>>,
+    #[cfg(any(target_os = "macos", target_os = "ios"))] on_web_content_process_terminate: Option<
+      Arc<OnWebContentProcessTerminate<R>>,
+    >,
     uri_scheme_protocols: HashMap<String, Arc<webview::UriSchemeProtocol<R>>>,
     state: StateManager,
     #[cfg(desktop)] menu_event_listener: Vec<crate::app::GlobalMenuEventListener<AppHandle<R>>>,
@@ -284,6 +290,8 @@ impl<R: Runtime> AppManager<R> {
         webviews: Mutex::default(),
         invoke_handler,
         on_page_load,
+        #[cfg(any(target_os = "macos", target_os = "ios"))]
+        on_web_content_process_terminate,
         uri_scheme_protocols: Mutex::new(uri_scheme_protocols),
         event_listeners: Arc::new(webview_event_listeners),
         invoke_initialization_script,
@@ -372,6 +380,7 @@ impl<R: Runtime> AppManager<R> {
     }
   }
 
+  // TODO: Change to return `crate::Result` here in v3
   pub fn get_asset(
     &self,
     mut path: String,
@@ -417,49 +426,42 @@ impl<R: Runtime> AppManager<R> {
         asset_path = fallback;
         asset
       })
-      .ok_or_else(|| crate::Error::AssetNotFound(path.clone()))
-      .map(Cow::into_owned);
+      .ok_or_else(|| {
+        let error = crate::Error::AssetNotFound(path.clone());
+        log::error!("{error}");
+        Box::new(error)
+      })?;
 
     let mut csp_header = None;
     let is_html = asset_path.as_ref().ends_with(".html");
 
-    match asset_response {
-      Ok(asset) => {
-        let final_data = if is_html {
-          let mut asset = String::from_utf8_lossy(&asset).into_owned();
-          if let Some(csp) = self.csp() {
-            #[allow(unused_mut)]
-            let mut csp_map = set_csp(&mut asset, &self.assets, &asset_path, self, csp);
-            #[cfg(feature = "isolation")]
-            if let Pattern::Isolation { schema, .. } = &*self.pattern {
-              let default_src = csp_map
-                .entry("default-src".into())
-                .or_insert_with(Default::default);
-              default_src.push(crate::pattern::format_real_schema(
-                schema,
-                _use_https_schema,
-              ));
-            }
+    let final_data = if is_html {
+      let mut asset = String::from_utf8_lossy(&asset_response).into_owned();
+      if let Some(csp) = self.csp() {
+        #[allow(unused_mut)]
+        let mut csp_map = set_csp(&mut asset, &self.assets, &asset_path, self, csp);
+        #[cfg(feature = "isolation")]
+        if let Pattern::Isolation { schema, .. } = &*self.pattern {
+          let default_src = csp_map.entry("default-src".to_owned()).or_default();
+          default_src.push(crate::pattern::format_real_schema(
+            schema,
+            _use_https_schema,
+          ));
+        }
 
-            csp_header.replace(Csp::DirectiveMap(csp_map).to_string());
-          }
+        csp_header.replace(Csp::DirectiveMap(csp_map).to_string());
+      }
 
-          asset.into_bytes()
-        } else {
-          asset
-        };
-        let mime_type = tauri_utils::mime_type::MimeType::parse(&final_data, &path);
-        Ok(Asset {
-          bytes: final_data,
-          mime_type,
-          csp_header,
-        })
-      }
-      Err(e) => {
-        log::error!("{:?}", e);
-        Err(Box::new(e))
-      }
-    }
+      asset.into_bytes()
+    } else {
+      asset_response.into_owned()
+    };
+    let mime_type = tauri_utils::mime_type::MimeType::parse(&final_data, &path);
+    Ok(Asset {
+      bytes: final_data,
+      mime_type,
+      csp_header,
+    })
   }
 
   pub(crate) fn listeners(&self) -> &Listeners {
@@ -761,6 +763,8 @@ mod test {
       context,
       PluginStore::default(),
       Box::new(|_| false),
+      None,
+      #[cfg(any(target_os = "macos", target_os = "ios"))]
       None,
       Default::default(),
       StateManager::new(),