+++ crates/tauri-runtime-wry/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## \[2.10.1]
+
+### Dependencies
+
+- Upgraded to `tauri-utils@2.8.3`
+- Upgraded to `tauri-runtime@2.10.1`
+
 ## \[2.10.0]
 
 ### Bug Fixes